[TIM-456] Simplified the RelationSearcher inherintence tree and removed unused methods.
